Renamed the package from "bot_engine" to "gramhopper"
