+++ gramhopper/responses/response_helper.py
