+++ gramhopper/trigger_result.py
