+++ gramhopper/triggers/event_streak_trigger.py
