+++ gramhopper/triggers/filter_triggers.py
