+++ gramhopper/triggers/text_triggers.py
